requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

# Data
datasets>=2.19.0
Pillow>=10.0.0
albumentations>=1.4.0
label-studio-sdk>=0.0.32

# Experiment Tracking
wandb>=0.16.0

# Serving
fastapi>=0.110.0
uvicorn>=0.29.0
gradio>=4.20.0
python-multipart>=0.0.9

# Optimization
onnx>=1.16.0
onnxruntime-gpu>=1.17.0

# Evaluation
jiwer>=3.0.0
nltk>=3.8.0
scikit-learn>=1.4.0

# Dev / Test
pytest>=8.0.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
